backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
pydantic==2.5.0
requests==2.31.0
psycopg2-binary==2.9.9
pydantic[email]==2.5.0